backend/requirements.txt
asgiref==3.8.1
astroid==3.3.10
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2024.8.30
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.3
daphne==4.2.0
dill==0.4.0
Django==5.1.2
django-background-tasks==1.2.8
django-cleanup==9.0.0
django-otp==1.5.4
django-vite==3.0.5
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inertia-django==0.6.0
isort==6.0.1
mccabe==0.7.0
msgpack==1.1.0
msgpack-python==0.5.6
numpy==2.3.2
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
psycopg==3.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pylint==3.3.7
pyOpenSSL==25.1.0
qrcode==8.0
redis==6.2.0
requests==2.32.3
ruff==0.12.5
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.1
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.2